+++ src/bin/dev/main.rs
@@ -144,7 +144,7 @@ fn gen_samples_entries(
                 autumnus::Options {
                     language: Some(file_name),
                     formatter: autumnus::FormatterOption::HtmlInline {
-                        theme: Some(theme),
+                        theme: Some(theme.clone()),
                         pre_class: Some(
                             "w-full overflow-auto rounded-lg p-8 font-mono text-sm antialiased leading-6",
                         ),